+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 All releases can be found on https://code.vikunja.io/vikunja/releases.
 
+## [Unreleased]
+
+### Added
+
+* Projects and teams now have opaque `uid` fields with ULIDs. Existing
+  databases are backfilled by migration `20250616230622`.
+
 ## [0.24.6] - 2024-12-22
 
 ### Bug Fixes